mounting conarydb optional; more conarydb tuning
Some tests are showing that mounting with barrier=0,data=writeback
gives closely equivalent performance overall to mounting a tmpfs
for conarydb.  For now, the tmpfs is optional.  It might be removed
for simplicity later.

Added default_cache_size for conarydb tuning.  It is deprecated (so
might not always be honored), but setting it now may make it easier
to benchmark to determine whether Conary could usefully set cache_size.