system call numbers chosen correctly for cross-build
Previously, if a build was being done under a 32-bit personality
with a 64-bit python, it would try to use the 32-bit syscall
numbers for clone and getpid, and thus fail.  Now it determines
whether the python interpreter in use is 32-bit or 64-bit as
well as using the personality to choose the system call numbers
to use.  This enables 32-bit builds on 64-bit systems using
the "setarch i686" command to choose the i686 personality.

Signed-off-by: Michael K Johnson <johnsonm@foresightlinux.org>

Fixes #28